printWidth=120

+++ core/backend/src/BriefcaseManager.ts
@@ -37,11 +37,7 @@ import {
 import { TelemetryEvent } from "@itwin/core-telemetry";
 import { AcquireNewBriefcaseIdArg } from "./BackendHubAccess";
 import { BackendLoggerCategory } from "./BackendLoggerCategory";
-import {
-  CheckpointManager,
-  CheckpointProps,
-  ProgressFunction,
-} from "./CheckpointManager";
+import { CheckpointManager, CheckpointProps, ProgressFunction } from "./CheckpointManager";
 import { BriefcaseDb, IModelDb, TokenArg } from "./IModelDb";
 import { IModelHost } from "./IModelHost";
 import { IModelJsFs } from "./IModelJsFs";
@@ -51,9 +47,7 @@ const loggerCategory = BackendLoggerCategory.IModelDb;
 /** The argument for [[BriefcaseManager.downloadBriefcase]]
  * @public
  */
-export interface RequestNewBriefcaseArg
-  extends TokenArg,
-    RequestNewBriefcaseProps {
+export interface RequestNewBriefcaseArg extends TokenArg, RequestNewBriefcaseProps {
   /** If present, a function called periodically during the download to indicate progress.
    * @note return non-zero from this function to abort the download.
    */
@@ -114,20 +108,12 @@ export class BriefcaseManager {
 
   /** @internal */
   public static getChangeCachePathName(iModelId: GuidString): LocalFileName {
-    return path.join(
-      this.getIModelPath(iModelId),
-      iModelId.concat(".bim.ecchanges")
-    );
+    return path.join(this.getIModelPath(iModelId), iModelId.concat(".bim.ecchanges"));
   }
 
   /** @internal */
-  public static getChangedElementsPathName(
-    iModelId: GuidString
-  ): LocalFileName {
-    return path.join(
-      this.getIModelPath(iModelId),
-      iModelId.concat(".bim.elems")
-    );
+  public static getChangedElementsPathName(iModelId: GuidString): LocalFileName {
+    return path.join(this.getIModelPath(iModelId), iModelId.concat(".bim.elems"));
   }
 
   private static _briefcaseSubDir = "briefcases";
@@ -142,10 +128,7 @@ export class BriefcaseManager {
    * @see getIModelPath
    */
   public static getFileName(briefcase: BriefcaseProps): LocalFileName {
-    return path.join(
-      this.getBriefcaseBasePath(briefcase.iModelId),
-      `${briefcase.briefcaseId}.bim`
-    );
+    return path.join(this.getBriefcaseBasePath(briefcase.iModelId), `${briefcase.briefcaseId}.bim`);
   }
 
   private static setupCacheDir(cacheRootDir: LocalDirName) {
@@ -175,18 +158,12 @@ export class BriefcaseManager {
    * iModels in the briefcase cache are returned.
    * @note usually there should only be one briefcase per iModel.
    */
-  public static getCachedBriefcases(
-    iModelId?: GuidString
-  ): LocalBriefcaseProps[] {
+  public static getCachedBriefcases(iModelId?: GuidString): LocalBriefcaseProps[] {
     const briefcaseList: LocalBriefcaseProps[] = [];
     const iModelDirs = IModelJsFs.readdirSync(this._cacheDir);
     for (const iModelDir of iModelDirs) {
       if (iModelId && iModelId !== iModelDir) continue;
-      const bcPath = path.join(
-        this._cacheDir,
-        iModelDir,
-        this._briefcaseSubDir
-      );
+      const bcPath = path.join(this._cacheDir, iModelDir, this._briefcaseSubDir);
       try {
         if (!IModelJsFs.lstatSync(bcPath)?.isDirectory) continue;
       } catch (err) {
@@ -199,10 +176,7 @@ export class BriefcaseManager {
           try {
             const fileName = path.join(bcPath, briefcaseName);
             const fileSize = IModelJsFs.lstatSync(fileName)?.size ?? 0;
-            const db = IModelDb.openDgnDb(
-              { path: fileName },
-              OpenMode.Readonly
-            );
+            const db = IModelDb.openDgnDb({ path: fileName }, OpenMode.Readonly);
             briefcaseList.push({
               fileName,
               iTwinId: db.getITwinId(),
@@ -229,18 +203,14 @@ export class BriefcaseManager {
    * @note this does check whether the id was actually acquired by the caller.
    */
   public static isValidBriefcaseId(id: BriefcaseId) {
-    return (
-      id >= BriefcaseIdValue.FirstValid && id <= BriefcaseIdValue.LastValid
-    );
+    return id >= BriefcaseIdValue.FirstValid && id <= BriefcaseIdValue.LastValid;
   }
 
   /** Acquire a new briefcaseId from iModelHub for the supplied iModelId
    * @note usually there should only be one briefcase per iModel per user. If a single user acquires more than one briefcaseId,
    * it's a good idea to supply different aliases for each of them.
    */
-  public static async acquireNewBriefcaseId(
-    arg: AcquireNewBriefcaseIdArg
-  ): Promise<BriefcaseId> {
+  public static async acquireNewBriefcaseId(arg: AcquireNewBriefcaseIdArg): Promise<BriefcaseId> {
     return IModelHost.hubAccess.acquireNewBriefcaseId(arg);
   }
 
@@ -264,18 +234,12 @@ export class BriefcaseManager {
    * for the IModelId.
    * @note *It is invalid to edit briefcases on a shared network drive* and that is a sure way to corrupt your briefcase (see https://www.sqlite.org/howtocorrupt.html)
    */
-  public static async downloadBriefcase(
-    arg: RequestNewBriefcaseArg
-  ): Promise<LocalBriefcaseProps> {
-    const briefcaseId =
-      arg.briefcaseId ?? (await this.acquireNewBriefcaseId(arg));
+  public static async downloadBriefcase(arg: RequestNewBriefcaseArg): Promise<LocalBriefcaseProps> {
+    const briefcaseId = arg.briefcaseId ?? (await this.acquireNewBriefcaseId(arg));
     const fileName = arg.fileName ?? this.getFileName({ ...arg, briefcaseId });
 
     if (IModelJsFs.existsSync(fileName))
-      throw new IModelError(
-        IModelStatus.FileAlreadyExists,
-        `Briefcase "${fileName}" already exists`
-      );
+      throw new IModelError(IModelStatus.FileAlreadyExists, `Briefcase "${fileName}" already exists`);
 
     const asOf = arg.asOf ?? IModelVersion.latest().toJSON();
     const changeset = await IModelHost.hubAccess.getChangesetFromVersion({
@@ -324,10 +288,7 @@ export class BriefcaseManager {
       nativeDb.enableWalMode(); // local briefcases should use WAL journal mode
       nativeDb.resetBriefcaseId(briefcaseId);
       if (nativeDb.getCurrentChangeset().id !== checkpoint.changeset.id)
-        throw new IModelError(
-          IModelStatus.InvalidId,
-          `Downloaded briefcase has wrong changesetId: ${fileName}`
-        );
+        throw new IModelError(IModelStatus.InvalidId, `Downloaded briefcase has wrong changesetId: ${fileName}`);
     } finally {
       nativeDb.saveChanges();
       nativeDb.closeIModel();
@@ -347,10 +308,7 @@ export class BriefcaseManager {
    * @note generally, this method should not be called directly. Instead use [[deleteBriefcaseFiles]].
    * @see deleteBriefcaseFiles
    */
-  public static async releaseBriefcase(
-    accessToken: AccessToken,
-    briefcase: BriefcaseProps
-  ): Promise<void> {
+  public static async releaseBriefcase(accessToken: AccessToken, briefcase: BriefcaseProps): Promise<void> {
     if (this.isValidBriefcaseId(briefcase.briefcaseId))
       return IModelHost.hubAccess.releaseBriefcase({
         accessToken,
@@ -366,10 +324,7 @@ export class BriefcaseManager {
    * @param filePath the full file name of the Briefcase to delete
    * @param accessToken for releasing the briefcaseId
    */
-  public static async deleteBriefcaseFiles(
-    filePath: LocalFileName,
-    accessToken?: AccessToken
-  ): Promise<void> {
+  public static async deleteBriefcaseFiles(filePath: LocalFileName, accessToken?: AccessToken): Promise<void> {
     try {
       const db = IModelDb.openDgnDb({ path: filePath }, OpenMode.Readonly);
       const briefcase: BriefcaseProps = {
@@ -389,10 +344,7 @@ export class BriefcaseManager {
     try {
       if (IModelJsFs.existsSync(filePath)) IModelJsFs.unlinkSync(filePath);
     } catch (err) {
-      throw new IModelError(
-        IModelStatus.BadRequest,
-        `cannot delete briefcase file ${err}`
-      );
+      throw new IModelError(IModelStatus.BadRequest, `cannot delete briefcase file ${err}`);
     }
 
     // next, delete all files that start with the briefcase's filePath (e.g. "a.bim-locks", "a.bim-journal", etc.)
@@ -401,8 +353,7 @@ export class BriefcaseManager {
       const fileName = path.basename(filePath);
       const files = IModelJsFs.readdirSync(dirName);
       for (const file of files) {
-        if (file.startsWith(fileName))
-          this.deleteFile(path.join(dirName, file)); // don't throw on error
+        if (file.startsWith(fileName)) this.deleteFile(path.join(dirName, file)); // don't throw on error
       }
     } catch (err) {}
   }
@@ -415,10 +366,7 @@ export class BriefcaseManager {
     try {
       IModelJsFs.unlinkSync(pathname);
     } catch (error) {
-      Logger.logError(
-        loggerCategory,
-        `Cannot delete file ${pathname}, ${error}`
-      );
+      Logger.logError(loggerCategory, `Cannot delete file ${pathname}, ${error}`);
       return false;
     }
     return true;
@@ -435,10 +383,7 @@ export class BriefcaseManager {
 
       IModelJsFs.rmdirSync(folderPathname);
     } catch (error) {
-      Logger.logError(
-        loggerCategory,
-        `Cannot delete folder: ${folderPathname}`
-      );
+      Logger.logError(loggerCategory, `Cannot delete folder: ${folderPathname}`);
       return false;
     }
     return true;
@@ -467,9 +412,7 @@ export class BriefcaseManager {
    *  - Does not throw any errors, but logs them.
    *  - returns true if the delete was successful.
    */
-  private static deleteFolderAndContents(
-    folderPathname: LocalDirName
-  ): boolean {
+  private static deleteFolderAndContents(folderPathname: LocalDirName): boolean {
     if (!IModelJsFs.existsSync(folderPathname)) return true;
 
     let status = false;
@@ -480,10 +423,7 @@ export class BriefcaseManager {
     return status;
   }
 
-  private static async applySingleChangeset(
-    db: IModelDb,
-    changesetFile: ChangesetFileProps
-  ) {
+  private static async applySingleChangeset(db: IModelDb, changesetFile: ChangesetFileProps) {
     if (changesetFile.changesType === ChangesetType.Schema) db.clearCaches(); // for schema changesets, statement caches may become invalid. Do this *before* applying, in case db needs to be closed (open statements hold db open.)
 
     db.nativeDb.applyChangeset(changesetFile);
@@ -494,16 +434,10 @@ export class BriefcaseManager {
   }
 
   /** @internal */
-  public static async pullAndApplyChangesets(
-    db: IModelDb,
-    arg: PullChangesArgs
-  ): Promise<void> {
+  public static async pullAndApplyChangesets(db: IModelDb, arg: PullChangesArgs): Promise<void> {
     if (!db.isOpen || db.nativeDb.isReadonly())
       // don't use db.isReadonly - we reopen the file writable just for this operation but db.isReadonly is still true
-      throw new IModelError(
-        ChangeSetStatus.ApplyError,
-        "Briefcase must be open ReadWrite to process change sets"
-      );
+      throw new IModelError(ChangeSetStatus.ApplyError, "Briefcase must be open ReadWrite to process change sets");
 
     let currentIndex = db.changeset.index;
     if (currentIndex === undefined)
@@ -533,18 +467,14 @@ export class BriefcaseManager {
 
     if (reverse) changesets.reverse();
 
-    for (const changeset of changesets)
-      await this.applySingleChangeset(db, changeset);
+    for (const changeset of changesets) await this.applySingleChangeset(db, changeset);
 
     // notify listeners
     db.notifyChangesetApplied();
   }
 
   /** create a changeset from the current changes, and push it to iModelHub */
-  private static async pushChanges(
-    db: BriefcaseDb,
-    arg: PushChangesArgs
-  ): Promise<void> {
+  private static async pushChanges(db: BriefcaseDb, arg: PushChangesArgs): Promise<void> {
     const changesetProps = db.nativeDb.startCreateChangeset();
     changesetProps.briefcaseId = db.briefcaseId;
     changesetProps.description = arg.description;
@@ -592,21 +522,14 @@ export class BriefcaseManager {
   /** Pull/merge (if necessary), then push all local changes as a changeset. Called by [[BriefcaseDb.pushChanges]]
    * @internal
    */
-  public static async pullMergePush(
-    db: BriefcaseDb,
-    arg: PushChangesArgs
-  ): Promise<void> {
+  public static async pullMergePush(db: BriefcaseDb, arg: PushChangesArgs): Promise<void> {
     let retryCount = arg.mergeRetryCount ?? 5;
     while (true) {
       try {
         await BriefcaseManager.pullAndApplyChangesets(db, arg);
         return await BriefcaseManager.pushChanges(db, arg);
       } catch (err: any) {
-        if (
-          retryCount-- <= 0 ||
-          err.errorNumber !== IModelHubStatus.PullIsRequired
-        )
-          throw err;
+        if (retryCount-- <= 0 || err.errorNumber !== IModelHubStatus.PullIsRequired) throw err;
         await (arg.mergeRetryDelay ?? BeDuration.fromSeconds(3)).wait();
       }
     }
